[SPARK-47534][SQL] Move `o.a.s.variant` to `o.a.s.types.variant`
