+++ common/variant/src/main/java/org/apache/spark/types/variant/Variant.java
@@ -15,11 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.spark.variant;
+package org.apache.spark.types.variant;
 
 /**
- * An exception indicating that we are attempting to build a variant with it value or metadata
- * exceeding the 16MiB size limit.
+ * This class is structurally equivalent to {@link org.apache.spark.unsafe.types.VariantVal}. We
+ * define a new class to avoid depending on or modifying Spark.
  */
-public class VariantSizeLimitException extends RuntimeException {
+public final class Variant {
+  private final byte[] value;
+  private final byte[] metadata;
+
+  public Variant(byte[] value, byte[] metadata) {
+    this.value = value;
+    this.metadata = metadata;
+  }
+
+  public byte[] getValue() {
+    return value;
+  }
+
+  public byte[] getMetadata() {
+    return metadata;
+  }
 }
